src/data/usecases/load-account-by-token/db-load-account-by-token.ts
import { IDecrypter } from '@/data/protocols/cryptography/decrypter'
import { ILoadAccountByTokenRepository } from '@/data/protocols/db/account/load-account-by-token-repository'
import { TAccountModel, ILoadAccountByToken } from '@/presentation/middlewares/auth-middleware-protocols'

export class DbLoadAccountByToken implements ILoadAccountByToken {
  constructor (
    private readonly decrypter: IDecrypter,
    private readonly loadAccountByTokenRepository: ILoadAccountByTokenRepository
  ) { }

  async load (accessToken: string, role?: string): Promise<TAccountModel> {
    const token = await this.decrypter.decrypt(accessToken)
    if (token) {
      const account = await this.loadAccountByTokenRepository.loadByToken(accessToken, role)
      if (account) {
        return account
      }
    }
    return null
  }
}
